Move and update tests to reflect changes in API
