+++ tests/test_benchmark_run.py
@@ -3,7 +3,8 @@ import json
 import testfixtures
 import logging
 
-from src.benchmark_run import SpecJBBRun, InvalidRunConfigurationException, JvmRunOptions, SpecJBBComponentOptions, SpecJBBComponentTypes
+from src.validate import JvmRunOptions
+from src.benchmark_run import SpecJBBRun, InvalidRunConfigurationException, SpecJBBComponentOptions, SpecJBBComponentTypes
 
 
 class TestBenchmarkRun(unittest.TestCase):
@@ -79,69 +80,6 @@ class TestBenchmarkRun(unittest.TestCase):
                 # need to have some actual logging output
                 self.assertTrue(l.actual())
 
-
-class TestJvmRunOptions(unittest.TestCase):
-    def test_given_none_will_fill_defaults(self):
-        self.assertTrue(JvmRunOptions())
-
-    def test_given_str(self):
-        java_path = "java"
-        j = JvmRunOptions(java_path)
-
-        self.assertEqual(j.path, java_path)
-        self.assertEqual(j["path"], java_path)
-        self.assertEqual(j.options, [])
-        self.assertEqual(j["options"], [])
-
-    def test_given_list(self):
-        java_list = ["java", "-jar", "example_jar"]
-        j = JvmRunOptions(java_list)
-
-        self.assertEqual(j.path, java_list[0])
-        self.assertEqual(j["path"], java_list[0])
-        self.assertEqual(j.options, java_list[1:])
-        self.assertEqual(j["options"], java_list[1:])
-
-    def test_given_dict(self):
-        valid = {
-            "path": "java",
-            "options": ["-jar", "example_jar"],
-        }
-
-        j = JvmRunOptions(valid)
-
-        self.assertEqual(j.path, valid["path"])
-        self.assertEqual(j["path"], valid["path"])
-        self.assertEqual(j.options, valid["options"])
-        self.assertEqual(j["options"], valid["options"])
-
-    def test_with_dict_missing_options(self):
-        valid = {
-            "path": "java",
-        }
-
-        j = JvmRunOptions(valid)
-
-        self.assertEqual(j.path, valid["path"])
-        self.assertEqual(j["path"], valid["path"])
-        self.assertEqual(j.options, [])
-        self.assertEqual(j["options"], [])
-
-    def test_validates_dictionaries(self):
-        invalid_missing_path = {
-            "options": []
-        }
-
-        with self.assertRaises(Exception):
-            j = JvmRunOptions(invalid_missing_path)
-
-        invalid_types = {
-            "path": 2,
-            "options": {}
-        }
-
-        with self.assertRaises(Exception):
-            j = JvmRunOptions(invalid_types)
 
 
 class TestSpecJBBComponentOptions(unittest.TestCase):
