+++ tests/test_validate.py
@@ -194,3 +194,57 @@ class TestJavaRunOptionsValidator(TestCase):
                 }
 
         self.assertEqual(JvmRunOptions.validate(valid), valid)
+
+    def test_given_none_will_fill_defaults(self):
+        self.assertTrue(JvmRunOptions.validate(None))
+
+    def test_given_str(self):
+        java_path = "java"
+        j = JvmRunOptions.validate(java_path)
+
+        self.assertEqual(j["path"], java_path)
+        self.assertEqual(j["options"], [])
+
+    def test_given_list(self):
+        java_list = ["java", "-jar", "example_jar"]
+        j = JvmRunOptions.validate(java_list)
+
+        self.assertEqual(j["path"], java_list[0])
+        self.assertEqual(j["options"], java_list[1:])
+
+    def test_given_dict(self):
+        valid = {
+            "path": "java",
+            "options": ["-jar", "example_jar"],
+        }
+
+        j = JvmRunOptions.validate(valid)
+
+        self.assertEqual(j["path"], valid["path"])
+        self.assertEqual(j["options"], valid["options"])
+
+    def test_with_dict_missing_options(self):
+        valid = {
+            "path": "java",
+        }
+
+        j = JvmRunOptions.validate(valid)
+
+        self.assertEqual(j["path"], valid["path"])
+        self.assertEqual(j["options"], [])
+
+    def test_validates_dictionaries(self):
+        invalid_missing_path = {
+            "options": []
+        }
+
+        with self.assertRaises(Exception):
+            JvmRunOptions.validate(invalid_missing_path)
+
+        invalid_types = {
+            "path": 2,
+            "options": {}
+        }
+
+        with self.assertRaises(Exception):
+            JvmRunOptions(invalid_types)
